refactor: 커서 추출 함수 helpers 디렉토리로 이동
