+++ src/domain/helpers/get-cursor.ts
